+++ relay/src/postgres.rs
@@ -1,5 +1,5 @@
 #![allow(clippy::cast_possible_truncation)]
-use crate::{Error, Job, JobId, Queue, Result};
+use crate::{Error, RawJob, Result};
 use chrono::{NaiveDateTime, TimeZone, Utc};
 use log::LevelFilter;
 use metrics::counter;
@@ -42,8 +42,8 @@ impl PgStore {
         let pool = PgPoolOptions::new()
             .max_connections(100)
             .min_connections(10)
-            .connect_timeout(Duration::from_secs(60))
-            .idle_timeout(Duration::from_secs(20))
+            .connect_timeout(Duration::from_secs(5))
+            .idle_timeout(Duration::from_secs(60))
             .after_connect(|conn| {
                 Box::pin(async move {
                     // Insurance as if not at least this isolation mode then some queries are not
@@ -89,7 +89,8 @@ impl PgStore {
     /// # Errors
     ///
     /// Will return `Err` if there is any communication issues with the backend Postgres DB.
-    pub async fn enqueue(&self, job: &Job) -> Result<()> {
+    #[tracing::instrument(name = "pg_enqueue", level = "debug", skip_all, fields(job_id=%job.id, queue=%job.queue))]
+    pub async fn enqueue(&self, job: &RawJob) -> Result<()> {
         let now = Utc::now();
         let run_at = if let Some(run_at) = job.run_at {
             run_at
@@ -123,12 +124,9 @@ impl PgStore {
                         }
                     }
                 }
-
-                Error::Postgres {
-                    message: e.to_string(),
-                    is_retryable: is_retryable(e),
-                }
+                e.into()
             })?;
+        debug!("enqueued job");
         Ok(())
     }
 
@@ -142,11 +140,9 @@ impl PgStore {
     /// # Errors
     ///
     /// Will return `Err` if there is any communication issues with the backend Postgres DB.
-    pub async fn enqueue_batch(&self, jobs: &[Job]) -> Result<()> {
-        let mut transaction = self.pool.begin().await.map_err(|e| Error::Postgres {
-            message: e.to_string(),
-            is_retryable: is_retryable(e),
-        })?;
+    #[tracing::instrument(name = "pg_enqueue_batch", level = "debug", skip_all, fields(jobs = jobs.len()))]
+    pub async fn enqueue_batch(&self, jobs: &[RawJob]) -> Result<()> {
+        let mut transaction = self.pool.begin().await?;
 
         for job in jobs {
             let now = Utc::now();
@@ -183,18 +179,11 @@ impl PgStore {
             .bind(&now)
             .bind(&run_at)
             .execute(&mut transaction)
-            .await
-            .map_err(|e| Error::Postgres {
-                message: e.to_string(),
-                is_retryable: is_retryable(e),
-            })?;
+            .await?;
         }
 
-        transaction.commit().await.map_err(|e| Error::Postgres {
-            message: e.to_string(),
-            is_retryable: is_retryable(e),
-        })?;
-
+        transaction.commit().await?;
+        debug!("enqueued batch");
         Ok(())
     }
 
@@ -203,17 +192,15 @@ impl PgStore {
     /// # Errors
     ///
     /// Will return `Err` if there is any communication issues with the backend Postgres DB.
-    pub async fn remove(&self, queue: &Queue, job_id: &JobId) -> Result<()> {
+    #[tracing::instrument(name = "pg_remove", level = "debug", skip_all, fields(job_id=%job_id, queue=%queue))]
+    pub async fn remove(&self, queue: &str, job_id: &str) -> Result<()> {
         sqlx::query("DELETE FROM jobs WHERE queue=$1 AND id=$2")
             .bind(queue)
             .bind(job_id)
             .execute(&self.pool)
-            .await
-            .map_err(|e| Error::Postgres {
-                message: e.to_string(),
-                is_retryable: is_retryable(e),
-            })?;
+            .await?;
 
+        debug!("removed job");
         Ok(())
     }
 
@@ -222,7 +209,8 @@ impl PgStore {
     /// # Errors
     ///
     /// Will return `Err` if there is any communication issues with the backend Postgres DB.
-    pub async fn next(&self, queue: &Queue, num_jobs: u32) -> Result<Option<Vec<Job>>> {
+    #[tracing::instrument(name = "pg_next", level = "debug", skip_all, fields(num_jobs=num_jobs, queue=%queue))]
+    pub async fn next(&self, queue: &str, num_jobs: u32) -> Result<Option<Vec<RawJob>>> {
         // MUST USE CTE WITH `FOR UPDATE SKIP LOCKED LIMIT` otherwise the Postgres Query Planner
         // CAN optimize the query which will cause MORE updates than the LIMIT specifies within
         // a nested loop.
@@ -269,7 +257,7 @@ impl PgStore {
             let timeout: PgInterval = row.get(2);
             let run_at: NaiveDateTime = row.get(6);
 
-            Job {
+            RawJob {
                 id: row.get(0),
                 queue: row.get(1),
                 timeout: (timeout.microseconds / 1_000_000) as i32,
@@ -282,15 +270,13 @@ impl PgStore {
             }
         })
         .fetch_all(&self.pool)
-        .await
-        .map_err(|e| Error::Postgres {
-            message: e.to_string(),
-            is_retryable: is_retryable(e),
-        })?;
+        .await?;
 
         if jobs.is_empty() {
+            debug!("fetched no jobs");
             Ok(None)
         } else {
+            debug!(fetched_jobs = jobs.len(), "fetched next job(s)");
             Ok(Some(jobs))
         }
     }
@@ -301,6 +287,7 @@ impl PgStore {
     ///
     /// Will return `Err` if there is any communication issues with the backend Postgres DB or the
     /// Job attempting to be updated cannot be found.
+    #[tracing::instrument(name = "pg_update", level = "debug", skip_all, fields(job_id=%job_id, queue=%queue))]
     pub async fn update(
         &self,
         queue: &str,
@@ -323,19 +310,17 @@ impl PgStore {
         .bind(job_id)
         .bind(state.map(|state| Some(Json(state))))
         .execute(&self.pool)
-        .await
-        .map_err(|e| Error::Postgres {
-            message: e.to_string(),
-            is_retryable: is_retryable(e),
-        })?
+        .await?
         .rows_affected();
 
         if rows_affected == 0 {
+            debug!("job not found");
             Err(Error::JobNotFound {
                 job_id: job_id.to_string(),
                 queue: queue.to_string(),
             })
         } else {
+            debug!("updated job");
             Ok(())
         }
     }
@@ -349,7 +334,8 @@ impl PgStore {
     /// # Errors
     ///
     /// Will return `Err` if there is any communication issues with the backend Postgres DB.
-    pub async fn reschedule(&self, job: &Job) -> Result<()> {
+    #[tracing::instrument(name = "pg_reschedule", level = "debug", skip_all, fields(job_id=%job.id, queue=%job.queue))]
+    pub async fn reschedule(&self, job: &RawJob) -> Result<()> {
         let now = Utc::now();
         let run_at = if let Some(run_at) = job.run_at {
             run_at
@@ -387,19 +373,17 @@ impl PgStore {
         .bind(&now)
         .bind(&run_at)
         .execute(&self.pool)
-        .await
-        .map_err(|e| Error::Postgres {
-            message: e.to_string(),
-            is_retryable: is_retryable(e),
-        })?
+        .await?
         .rows_affected();
 
         if rows_affected == 0 {
+            debug!("job not found");
             Err(Error::JobNotFound {
                 job_id: job.id.to_string(),
                 queue: job.queue.to_string(),
             })
         } else {
+            debug!("rescheduled job");
             Ok(())
         }
     }
@@ -409,6 +393,7 @@ impl PgStore {
     /// # Errors
     ///
     /// Will return `Err` if there is any communication issues with the backend Postgres DB.
+    #[tracing::instrument(name = "pg_reap_timeouts", level = "debug", skip(self))]
     pub async fn reap_timeouts(&self, interval_seconds: i64) -> Result<()> {
         let rows_affected = sqlx::query(
             r#"
@@ -418,11 +403,7 @@ impl PgStore {
         )
         .bind(interval_seconds)
         .execute(&self.pool)
-        .await
-        .map_err(|e| Error::Postgres {
-            message: e.to_string(),
-            is_retryable: is_retryable(e),
-        })?
+        .await?
         .rows_affected();
 
         // another instance has already updated OR time hasn't been hit yet
@@ -430,7 +411,7 @@ impl PgStore {
             return Ok(());
         }
 
-        debug!("running timeout reaper");
+        debug!("running timeout & delete reaper");
 
         let results: Vec<(String, i64)> = sqlx::query_as::<_, (String, i64)>(
             r#"
@@ -450,13 +431,10 @@ impl PgStore {
             "#,
         )
         .fetch_all(&self.pool)
-        .await
-        .map_err(|e| Error::Postgres {
-            message: e.to_string(),
-            is_retryable: is_retryable(e),
-        })?;
+        .await?;
 
         for (queue, count) in results {
+            debug!(queue = %queue, count = count, "retrying jobs");
             counter!("retries", u64::try_from(count).unwrap_or_default(), "queue" => queue);
         }
 
@@ -476,16 +454,13 @@ impl PgStore {
             "#,
         )
         .fetch_all(&self.pool)
-        .await
-        .map_err(|e| Error::Postgres {
-            message: e.to_string(),
-            is_retryable: is_retryable(e),
-        })?;
+        .await?;
 
         for (queue, count) in results {
             warn!(
-                "deleted {} records from queue '{}' that reached their max retries",
-                count, queue
+                count = count,
+                queue = %queue,
+                "deleted records from queue that reached their max retries"
             );
             counter!("errors", u64::try_from(count).unwrap_or_default(), "queue" => queue);
         }
@@ -525,5 +500,14 @@ fn is_retryable(e: SQLXError) -> bool {
                 | ErrorKind::UnexpectedEof
         ),
         _ => false,
+    }
+}
+
+impl From<sqlx::Error> for Error {
+    fn from(e: sqlx::Error) -> Self {
+        Error::Postgres {
+            message: e.to_string(),
+            is_retryable: is_retryable(e),
+        }
     }
 }